Initial commit of simple trace logging sample starter
This is the first commit of a sample starter that provides basic trace
logging functionality to spring boot apps.  This may not stay long term,
but if it does, modifications will have to be made to make it more
generic.   Currently, only methods annotated with the @Trace annotation
will be logged.

This will probably also not be the final structure that we choose to go
forward with, but it can be tweaked until we're happy.