#6 Refactored after comments and adding class level tracing via annotation
I've refactored based on comments in commit 14d62a3.  The @Trace
annotation can now be added at class level, causing all public methods
in that class to be traced.

NOTE: I haven't added a Conditional annotation to disable the trace
logging configuration (for added performance) currently as I would
rather find a way to not have to explicitly have a ConditionalOnProperty
annotation, but rather delegate to TraceLoggerConfig.isEnabled(), but I
need to figure out how to achieve this!