#16 #17 Adding configuration of messages and ability to disable trace logging
The trace-logger.message.entry/exit/exception messages can be configured
to override the default log messages.  The trace-logger.enabled property
can be used to disable trace logging for an application (Note that trace
logging is enabled unless this property is explicitly set to false ie.
if the property doesn't exist, then trace logging is enabled)