#18 #19 #20 Changed registrar strategy, moved shared code to core and fixed multi trace bug
I've added the ability to be able to annotate factory classes with
@SettingBackedBeanFactory.  Methods within these classes can be
annotated with @SettingBackedBean, specifying the setting that they are
to be backed by.  For every instance of this setting, a the factory
method will be called, passing the instance of the setting, and a new
bean is then subsequently created for each.

Note, I'm on holiday from Saturday so wanted to get this in before then,
but the core stuff still needs some tidy up, LOTS of unit tests and
LOADS of documentation.  Also it could be more customizable.

I kind of just went ahead and implemented this, but if we think its not
the best approach/not needed, then I'm happy for it to be reverted (it
was a good learning exercise for me at least!)